requirements.txt
pydantic~=2.8.2
rich~=13.7.1
names~=0.3.0
python-dotenv~=1.0.1
aiohttp~=3.9.5